refactor(NODE-3675): rename test / reorganize describe blocks
